requirements.txt
python-telegram-bot==20.5
python-dotenv==1.0.0
pandas==2.1.4
python-dateutil==2.8.2  # Зависимость для pandas
numpy==1.26.2  # Зависимость для pandas
loguru==0.7.2  # Для удобного логирования